feat: 定时任务，数据入库

+++ controller/hot.js
@@ -8,8 +8,10 @@ const cheerio = require("cheerio");
 const { comment } = require("./comment.js");
 const { downloadImg } = require("./utils.js");
 const isUrl = require("is-url");
+const dayjs = require("dayjs");
+const { insertIntoDB, connection } = require("./db.js");
 
-const hot = (apiUrl, dataItem, urlList, index, fileName, callback) => {
+const hot = (apiUrl, dataItem, urlList, index, fileName, callback, file) => {
     const id = dataItem.url.replace('https://www.zhihu.com/question/', '');
     console.log(`===================== 当前爬取热门话题：${dataItem.title} id: ${id} 中... =====================`);
     const hotUrl = apiUrl ?? `https://www.zhihu.com/api/v4/questions/${id}/feeds?include=data%5B%2A%5D.is_normal%2Cadmin_closed_comment%2Creward_info%2Cis_collapsed%2Cannotation_action%2Cannotation_detail%2Ccollapse_reason%2Cis_sticky%2Ccollapsed_by%2Csuggest_edit%2Ccomment_count%2Ccan_comment%2Ccontent%2Ceditable_content%2Cattachment%2Cvoteup_count%2Creshipment_settings%2Ccomment_permission%2Ccreated_time%2Cupdated_time%2Creview_info%2Crelevant_info%2Cquestion%2Cexcerpt%2Cis_labeled%2Cpaid_info%2Cpaid_info_content%2Creaction_instruction%2Crelationship.is_authorized%2Cis_author%2Cvoting%2Cis_thanked%2Cis_nothelp%3Bdata%5B%2A%5D.mark_infos%5B%2A%5D.url%3Bdata%5B%2A%5D.author.follower_count%2Cvip_info%2Cbadge%5B%2A%5D.topics%3Bdata%5B%2A%5D.settings.table_of_content.enabled&limit=5&offset=0&order=default&platform=desktop`;
@@ -29,22 +31,22 @@ const hot = (apiUrl, dataItem, urlList, index, fileName, callback) => {
 
     http(hotUrl, { headers: headers }).then((res) => {
         if (!res.data.paging.is_end) {
-            dealHotItem(res.data.data[0], res.data.data, 0, dataItem, index, res.data.paging.next, urlList, fileName, callback);
+            dealHotItem(res.data.data[0], res.data.data, 0, dataItem, index, res.data.paging.next, urlList, fileName, callback, file);
         } else {
             console.log(`===================== 热门话题：${dataItem.title} id: ${id} 处理完成... =====================`);
             const remainArr = urlList.slice(index + 1)
-            fs.writeFile(fileName, JSON.stringify(remainArr), 'utf8', (err) => {
+            fileName && fs.writeFile(fileName, JSON.stringify(remainArr), 'utf8', (err) => {
                 if (err) throw err;
                 console.log('删除源文件中的热门话题：' + dataItem.title);
             })
             if (index < urlList.length - 1) {
-                hot(undefined, urlList[index + 1], urlList, index + 1, fileName, callback);
+                hot(undefined, urlList[index + 1], urlList, index + 1, fileName, callback, file);
             } else {
-                console.log(`===================== ${fileName} 所有热门话题都爬完了 =====================`);
+                console.log(`===================== ${fileName ?? dayjs(file).toString()} 所有热门话题都爬完了 =====================`);
                 // 正常爬完删除源文件
-                fs.unlink(fileName, (err) => {
+                fileName && fs.unlink(fileName, (err) => {
                     if (err) throw err;
-                    callback();
+                    callback && callback();
                 })
             }
         }
@@ -53,17 +55,17 @@ const hot = (apiUrl, dataItem, urlList, index, fileName, callback) => {
         console.log("hot.js", err);
 
         if (index < urlList.length - 1) {
-            hot(undefined, urlList[index + 1], urlList, index + 1, fileName, callback);
+            hot(undefined, urlList[index + 1], urlList, index + 1, fileName, callback, file);
         } else {
-            console.log(`===================== ${fileName} 所有热门话题都爬完了 =====================`);
-            callback();
+            console.log(`===================== ${fileName ?? dayjs(file).toString()} 所有热门话题都爬完了 =====================`);
+            callback && callback();
         }
     })
 };
 
-const dealHotItem = (item, resArr, index, dataItem, pIndex, apiUrl, urlList, fileName, callback) => {
+const dealHotItem = (item, resArr, index, dataItem, pIndex, apiUrl, urlList, fileName, callback, file) => {
     const id = dataItem.url.replace('https://www.zhihu.com/question/', '');
-    const fileDir = path.join('/fedataset/zhihu', 'hot', `question_${id}`, String(item.target.id));
+    const fileDir = path.join('/fedataset/zhihu', 'hot-topic', `${file ? file.replace('.json', '') : dayjs().format('YYYY-MM-DD')}`, `question_${id}`, String(item.target.id));
     // fs.access(fileDir, fs.constants.R_OK, (err) => {
     // if (err) {
     // 文件夹不存在
@@ -106,15 +108,21 @@ const dealHotItem = (item, resArr, index, dataItem, pIndex, apiUrl, urlList, fil
                     const dealFilePath = path.join(fileDir, "deal.txt");
                     fs.writeFile(dealFilePath, str ?? "", (err) => {
                         if (err) throw err;
-                        console.log(item.target.id + '图片、源文件、处理文件写入完成');
+                        console.log(item.target.id + '图片、源文件、处理文件写入完成,开始写入数据库...');
+
+                        // 写入数据库
+                        const insertSQL = `insert into ${file ? `hot_${dayjs(file.replace('.json', '')).get("year")}_${dayjs(file.replace('.json', '')).get("month") + 1}` : `hot_${dayjs().get('year')}_${dayjs().get("month") + 1}`}(id,q_id,content,q_time,create_time,interface_data,path) values (?,?,?,?,?,?,?)`;
+                        const dataArr = [item.target.id, id, item.target.content ?? '', `${file ? file.replace('.json', '') : dayjs().format('YYYY-MM-DD')}`, dayjs().format('YYYY-MM-DD HH:mm:ss'), JSON.stringify(item), fileDir];
+
+                        insertIntoDB(insertSQL, dataArr);
 
                         if (index < resArr.length - 1) {
-                            dealHotItem(resArr[index + 1], resArr, index + 1, dataItem, pIndex, apiUrl, urlList, fileName, callback);
+                            dealHotItem(resArr[index + 1], resArr, index + 1, dataItem, pIndex, apiUrl, urlList, fileName, callback, file);
                         } else {
                             console.log('当前5条热门话题处理完成，约2s后开始获取下5条热门话题!');
                             setTimeout(() => {
-                                hot(apiUrl, dataItem, urlList, pIndex, fileName, callback);
-                            }, Math.random() * 2000);
+                                hot(apiUrl, dataItem, urlList, pIndex, fileName, callback, file);
+                            }, Math.random() * 2000)
                         }
                     });
                 };
